Don't throw error from _is_a_lwp on user process end during accont_gathe
This happens when user job executes many short processes. The race
condition between processes triggered by job and account_gather is
inevitable and we shouldn't throw error message to users in those cases.
If open of /proc/PID/status fails because of other reason, for instance
permission denied it should stay as error since in this case we know
that process exists but we were unable to account it into job.