EnforcePartLimits refering to nodes and partition configuration
Changing the behavior of EnforcePartLimits to apply to nodes
configuration and partition limits. For instance min-node-memory
requirements. Previously resources/nodes configuration was
required in any of the partitions the job was submitted to.

Changing the default value from NO to ANY, which effectively
results in partition limits being required by default(change)
and keeping the same default for nodes configuration.

Bug 7248.