fix --gpu-bind=map_gpu reuse map if #tasks > #map_elements
Documentation states that the list provided to map_gpu and mask_gpu,
will be reused  if needed if the number of tasks(ranks) is greater than
number of elements(including multiplication) is too small. Previously
the function returned NULL which was resulting in no
CUDA_VISIBLE_DEVICES affinity setting for tasks above number of
map elements.